[PATCH] Some i386 cleanups - MTRR, bootflag
This does:
 - fix one warning in bootflag.c
 - change a few longs to int and int to long in the MTRR driver
   to make it 64bit clean (should be a NOP for 32bit i386, but is needed
   for x86-64)
 - Convert the MTRR /proc interface to seq_file and remove the broken
   compute_ascii() hack.  This fixes some broken code e.g.  the old
   mtrr_write was completely broken because the loop checking for
   commands started with a "continue" - remove duplicated mtrr type
   strings.

+++ arch/i386/kernel/cpu/mtrr/amd.c
@@ -7,7 +7,7 @@
 
 static void
 amd_get_mtrr(unsigned int reg, unsigned long *base,
-	     unsigned long *size, mtrr_type * type)
+	     unsigned int *size, mtrr_type * type)
 {
 	unsigned long low, high;
